feat(event-logging): implement new event logging format (wip)
